fix: Phase 4 — infrastructure hardening (tee path, sqlite, runner/config, doctor) (#103)
* fix: tee hint prints full log path so agents can open it


* fix: sqlite schema includes success column; per-process retention cleanup; long totals

- Add `success` to CREATE TABLE so fresh DBs never migrate; wrap the legacy
  ALTER in a duplicate-column guard so two processes racing on one new file
  can't collide.
- Replace the 50-insert cleanup counter (which could never fire in a one-shot
  CLI) with a single retention purge at initialization.
- Store timestamps in UTC so the string range query is correct across offsets.
- Widen GainSummary/CommandGainDetail token totals to long and read sums via
  GetInt64 so long histories can't overflow int.
- Bound GetHistoryAsync with an internal LIMIT (500) so a huge tracking DB
  cannot load an unbounded result set. (No `gain --history` command exists to
  thread a caller-supplied limit through, so this is applied in the query.)


* fix: stdin EOF for captured children, atomic config save, robust kill/catch paths

- Redirect and immediately close stdin on captured child processes so a child
  that reads stdin sees EOF and exits instead of hanging.
- Widen KillProcess to also swallow Win32Exception/AggregateException from the
  kill race, not just InvalidOperationException.
- Write config to a temp file then atomically File.Move it into place so a
  crash mid-write cannot leave a truncated config.
- Scope the config-load and tee catch-alls to `when (ex is not
  OperationCanceledException)` so cancellation propagates instead of being
  silently swallowed.
- Build the SQLite connection string via SqliteConnectionStringBuilder so
  paths with special characters are escaped; route the DTK_DB_PATH /
  DTK_CONFIG_PATH / DTK_TEE_DIR overrides through one trimming helper.
- Guard ParseEnum against an empty value so it throws a friendly
  FormatException instead of IndexOutOfRangeException.


* fix: doctor shares real default paths and stops false-alarming fresh installs

- DoctorCommand now resolves the db/tee paths through the same shared statics
  the DI-registered tracker and tee service use (SqliteTracker.GetDefaultDbPath,
  FileTeeService.GetDefaultTeeDir) and the same DTK_DB_PATH override helper,
  killing the triplicated path logic and any drift between reported and actual.
- Treat a missing tracking-database directory as "will be created" (pass), the
  same as the tee-directory check, instead of failing a brand-new install.


* fix: honor DTK_TEE_DIR in doctor; clean up config temp file on save failure

Addresses Copilot review on #103:
- DoctorCommand now resolves the tee directory through DTK_TEE_DIR before
  config/default, matching FileTeeService's precedence, so doctor never reports
  a tee path different from the one dtk uses at runtime.
- JsonConfigProvider.SaveAsync clears the temp file in a finally so a failed or
  cancelled write/move can't leave a stray *.tmp behind.


---------

+++ src/DotnetTokenKiller.Domain/Tracking/CommandGainDetail.cs
@@ -10,9 +10,9 @@ namespace DotnetTokenKiller.Domain.Tracking;
 /// <param name="FailureDetail">Detail for runs that exited with a non-zero code, or <see langword="null"/> if none.</param>
 public sealed record CommandGainDetail(
     int RunCount,
-    int TotalInputTokens,
-    int TotalOutputTokens,
-    int TotalSavedTokens,
+    long TotalInputTokens,
+    long TotalOutputTokens,
+    long TotalSavedTokens,
     double AverageSavingsPercentage,
     CommandGainDetail? SuccessDetail = null,
     CommandGainDetail? FailureDetail = null);
